+++ src/Stormpath.Owin.Middleware/Route/ForgotPasswordRoute.cs
@@ -17,10 +17,11 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
 using Stormpath.Owin.Abstractions;
 using Stormpath.Owin.Abstractions.Configuration;
-using Stormpath.Owin.Abstractions.ViewModel;
 using Stormpath.Owin.Middleware.Internal;
+using Stormpath.Owin.Middleware.Model;
 using Stormpath.Owin.Middleware.ViewModelBuilder;
 
 namespace Stormpath.Owin.Middleware.Route
@@ -29,8 +30,6 @@ namespace Stormpath.Owin.Middleware.Route
     {
         public static bool ShouldBeEnabled(IntegrationConfiguration configuration)
             => configuration.Web.ForgotPassword.Enabled == true;
-                //|| (configuration.Web.ForgotPassword.Enabled == null && configuration.Tenant.PasswordResetWorkflowEnabled);
-                // TODO - any reason this needs to be dynamic now?
 
         protected override async Task<bool> GetHtmlAsync(IOwinEnvironment context, CancellationToken cancellationToken)
         {
@@ -43,57 +42,51 @@ namespace Stormpath.Owin.Middleware.Route
             return true;
         }
 
-        protected override Task<bool> PostHtmlAsync(IOwinEnvironment context, ContentType bodyContentType, CancellationToken cancellationToken)
+        protected override async Task<bool> PostHtmlAsync(IOwinEnvironment context, ContentType bodyContentType, CancellationToken cancellationToken)
         {
-            // todo how does the password reset flow work?
-            throw new Exception("TODO");
+            try
+            {
+                var body = await context.Request.GetBodyAsStringAsync(cancellationToken);
+                var model = PostBodyParser.ToModel<ForgotPasswordPostModel>(body, bodyContentType, _logger);
+                var formData = FormContentParser.Parse(body, _logger);
 
-            //try
-            //{
-            //    var body = await context.Request.GetBodyAsStringAsync(cancellationToken);
-            //    var model = PostBodyParser.ToModel<ForgotPasswordPostModel>(body, bodyContentType, _logger);
-            //    var formData = FormContentParser.Parse(body, _logger);
+                var stateToken = formData.GetString(StringConstants.StateTokenName);
+                var parsedStateToken = new StateTokenParser(_configuration.Okta.Application.Id, _configuration.OktaEnvironment.ClientSecret, stateToken, _logger);
+                if (!parsedStateToken.Valid)
+                {
+                    var queryString = QueryStringParser.Parse(context.Request.QueryString, _logger);
+                    var viewModelBuilder = new ForgotPasswordFormViewModelBuilder(_configuration, queryString);
+                    var viewModel = viewModelBuilder.Build();
+                    viewModel.Errors.Add("An error occurred. Please try again.");
 
-            //    var stateToken = formData.GetString(StringConstants.StateTokenName);
-            //    var parsedStateToken = new StateTokenParser(client, _configuration.Client.ApiKey, stateToken, _logger);
-            //    if (!parsedStateToken.Valid)
-            //    {
-            //        var queryString = QueryStringParser.Parse(context.Request.QueryString, _logger);
-            //        var viewModelBuilder = new ForgotPasswordFormViewModelBuilder(client, _configuration, queryString);
-            //        var viewModel = viewModelBuilder.Build();
-            //        viewModel.Errors.Add("An error occurred. Please try again.");
+                    await RenderViewAsync(context, _configuration.Web.ForgotPassword.View, viewModel, cancellationToken);
+                    return true;
+                }
 
-            //        await RenderViewAsync(context, _configuration.Web.ForgotPassword.View, viewModel, cancellationToken);
-            //        return true;
-            //    }
+                await _oktaClient.SendPasswordResetEmailAsync(model.Email, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(1002, ex, "ForgotRoute.PostHtml");
+            }
 
-            //    await application.SendPasswordResetEmailAsync(model.Email, cancellationToken);
-            //}
-            //catch (Exception ex)
-            //{
-            //    _logger.LogError(1002, ex, "ForgotRoute.PostHtml");
-            //}
-
-            //return await HttpResponse.Redirect(context, _configuration.Web.ForgotPassword.NextUri);
+            return await HttpResponse.Redirect(context, _configuration.Web.ForgotPassword.NextUri);
         }
 
-        protected override Task<bool> PostJsonAsync(IOwinEnvironment context, ContentType bodyContentType, CancellationToken cancellationToken)
+        protected override async Task<bool> PostJsonAsync(IOwinEnvironment context, ContentType bodyContentType, CancellationToken cancellationToken)
         {
-            // todo how does the password reset flow work?
-            throw new Exception("TODO");
+            try
+            {
+                var model = await PostBodyParser.ToModel<ForgotPasswordPostModel>(context, bodyContentType, _logger, cancellationToken);
 
-            //try
-            //{
-            //    var model = await PostBodyParser.ToModel<ForgotPasswordPostModel>(context, bodyContentType, _logger, cancellationToken);
+                await _oktaClient.SendPasswordResetEmailAsync(model.Email, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(1003, ex, "ForgotRoute.PostJson");
+            }
 
-            //    await application.SendPasswordResetEmailAsync(model.Email, cancellationToken);
-            //}
-            //catch(Exception ex)
-            //{
-            //    _logger.LogError(1003, ex, "ForgotRoute.PostJson");
-            //}
-
-            //return await JsonResponse.Ok(context);
+            return await JsonResponse.Ok(context);
         }
     }
 }