+++ changelog.md
@@ -12,3 +12,6 @@ Migration information TODO.
 	* CustomData _todo_
 * The `StateTokenBuilder` and `StateTokenParser` classes were moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.
 * The `*FormViewModelBuilder` classes was moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.
+* The `/forgot` and `/change` routes are now disabled by default. Previously, they would follow the Stormpath Directory configuration by default. The routes can be enabled or disabled by changing the `okta.web.forgotPassword.enabled` or `okta.web.changePassword.enabled` settings.
+* The `web.changePassword.errorUri` setting has been deprecated. Any errors that occur during password reset are rendered as error messages.
+* The shape of the user data passed to the `PreChangePasswordHandler` and `PostChangePasswordHandler` has changed. It is now an Okta profile map, as seen in this example: http://developer.okta.com/docs/api/resources/authn.html#verify-recovery-token